Move N_PREFIX to the start of PATH
Yarn requires Node as a dependency so we will be forced to install it with Homebrew. But we want to use the version(s) installed with n. Putting N_PREFIX at the start of the PATH means we can prefer items installed via NPM over Homebrew.

Also installed Yarn (though I probably don't need it).