Add a variable for NPM similar to Node
I haven't looked at this for a while (and forgot to commit it) but I think this is what's going on.

This would do a node install but then the npm installs could go to a different place because it was finding an older command.

I'm also moving away from global packages in favor of npx, though trash-cli is still a good global one because of my rm alias.